Add sophos utm template
We needed to add exports to sophos UTM template to allow our Active
directory template to reference the VPC. A PR[1] was made to upstream
project.  We add the template here for now but may change to reference
upstream later.  This template is forked from ver 9.506 [2].

The sophos template is larger than 51K so we also need to change the
deployment to upload templates to S3 bucket before validating them.

[1] https://github.com/sophos-iaas/aws-cf-templates/pull/21
[2] https://github.com/sophos-iaas/aws-cf-templates/blob/master/utm/9.506/autoscaling.template