Lambda to cleanup jumpcloud systems
A resource (i.e. EC2 instance) is provisioned with a CF stack and it
gets registered as a "system" in jumpcloud.  When the resource is removed
from AWS (via a stack delete) it will continue to be registered on jumpcloud.

We add this lambda to periodically clean up those registered systems on
jumpcloud.