Refactor to account for default parameters
Default parameters allows the user to optionally specify them in
sceptre.  If the parameter is not specified in a sceptre stack config
file reading the parameter in the hook will fail.  We need to
read the parameters from CF because that will always contain the
default value if it's not explicitly set in a sceptre config.